+++ stdnet/orm/fields.py
@@ -382,7 +382,7 @@ a :class:`datetime.date` instance.'''
         
         
 class DateTimeField(DateField):
-    '''An date :class:`AtomField` represented in Python by
+    '''A date :class:`AtomField` represented in Python by
 a :class:`datetime.datetime` instance.'''
     type = 'datetime'
     index = False